--- src/cesr/core/mod.rs
@@ -1,19 +1,0 @@
-pub mod bexter;
-pub mod cigar;
-pub mod common;
-pub mod counter;
-pub mod creder;
-pub mod dater;
-pub mod diger;
-pub mod indexer;
-pub mod matter;
-pub mod number;
-pub mod sadder;
-pub mod saider;
-pub mod salter;
-pub mod seqner;
-pub mod serder;
-pub mod siger;
-pub mod tholder;
-pub mod util;
-pub mod verfer;
