+++ src/cesr/mod.rs
@@ -1,23 +1,40 @@
-pub(crate) mod core;
-pub(crate) mod crypto;
+pub mod bexter;
+pub mod cigar;
+pub mod common;
+pub mod counter;
+pub mod creder;
+pub mod dater;
+pub mod diger;
+pub mod indexer;
+pub mod matter;
+pub mod number;
+pub mod sadder;
+pub mod saider;
+pub mod salter;
+pub mod seqner;
+pub mod serder;
+pub mod siger;
+pub mod tholder;
+pub mod util;
+pub mod verfer;
 
-pub use crate::cesr::core::{
-    bexter::{Bext, Bexter},
-    cigar::Cigar,
-    common,
-    counter::{tables as counter, Counter}, // This seems like it shoudl be an abstract class
-    creder::Creder,
-    dater::Dater,
-    diger::Diger,
-    indexer::{tables as indexer, Indexer},
-    matter::{tables as matter, Matter},
-    number::{tables as number, Number},
-    sadder::Sadder,
-    saider::Saider,
-    salter::Salter,
-    seqner::Seqner,
-    serder::Serder,
-    siger::Siger,
-    tholder::Tholder,
-    verfer::Verfer,
-};
+// pub use crate::cesr::{
+//     bexter::{Bext, Bexter},
+//     cigar::Cigar,
+//     common,
+//     counter::{tables as counter, Counter}, // This seems like it shoudl be an abstract class
+//     creder::Creder,
+//     dater::Dater,
+//     diger::Diger,
+//     indexer::{tables as indexer, Indexer},
+//     matter::{tables as matter, Matter},
+//     number::{tables as number, Number},
+//     sadder::Sadder,
+//     saider::Saider,
+//     salter::Salter,
+//     seqner::Seqner,
+//     serder::Serder,
+//     siger::Siger,
+//     tholder::Tholder,
+//     verfer::Verfer,
+// };
